+++ scenarios/flat_cgm_trace.json
